Update testcase for LessonName and StudentName
